Update viralrecon_wrapper.py
Viralrecon version 2.5 changed the sample naming convention of the "taxon" column for the pangolin files by replacing the underscore between the sample ID and reference genome to a space. Updated the 'get_pangolin_data' function to reflect this. Also added the addition of a backslash to the end of the outdirectory prefix variable if the outdirectory specified by the user did not have one. This fixes an issue observed in testing and correctly generates a new directory for the wrapper summary report.